infra/backend_services/requirements.txt
fastmcp
asyncpg
python-dotenv
azure-cosmos